Expanded plugin constructor capabilities (#481)
Support more constructors for plugins by adding the DataPrepperPluginConstructor annotation. This is the preferred constructor. If no other constructor is available for a plugin, use the no-op constructor. Updated the HTTPSource plugin to use this capability to receive both a configuration model and PluginMetrics via the constructor. For a single parameter, un-annotated constructor in plugins, the only supported parameter is once again PluginSetting.

Signed-off-by: David Venable <dlv@amazon.com>
Signed-off-by: Arun Lakshmanan <arnlaksh@amazon.com>