Upload and publish the JUnit test reports for some tests (#1336)
Upload and publish the JUnit test reports for the Gradle tests so that it is easier to track down issues. Additionally, build the Gradle GitHub Action for all pushes since this is the core build for the project. Renamed the Gradle build to be more compact and easier to find in the list of runs. Upload and publish JUnit reports for OpenSearch sink integration tests. Updated the Developer Guide with some information on Data Prepper continuous integration, including information on how to find unit test results.

Signed-off-by: David Venable <dlv@amazon.com>
Signed-off-by: Arunachalam Lakshmanan <arunachalam.l@gmail.com>