Support building on JDK 17 (#1430)
Provided a work-around suggested by the Spotless team to allow it to run successfully on JDK 17. This change allows developers to build Data Prepper using JDK 17. Updated the Spotless Gradle plugin, which also allows us to remove an older work-around related to cleaning the project root build directory. Updated the developer guide to clarify that Data Prepper can build with either JDK 11 or 17. Run the Gradle build and performance test builds on JDK 11 and 17 as part of the GitHub Actions CI.

Signed-off-by: David Venable <dlv@amazon.com>