Use the Region supplied in the S3 configuration for the STS client (#1527)
Use the Region supplied in the STS client so that users don't have to redefine the region multiple times. Renamed the AWS properties by removing the "aws_" prefix. Per PR discussions, allow the AWS region to be null and use a region as supplied by the SDK (often via the environment variable). Performed some refactoring to help with this.

Signed-off-by: David Venable <dlv@amazon.com>
Signed-off-by: Arun Lakshmanan <arnlaksh@amazon.com>