Add new method to parse files in InputCodec.
The InputCodec currently only supports reading from an InputStream. This
interface does not work for inputs that require random read access. In
order to support random access reads in addition to sequential reads,
the InputCode interface should support reading from an InputFile. For
convenience, the existing InputFile interface defined in the
parquet-common package is being used.

Signed-off-by: Adi Suresh <adsuresh@amazon.com>