Correct Parquet support for the S3 sink and a new multipart buffer type (#3186)
Correct the ParquetOutputCodec and moved into the S3 sink project for now. It has a few corrections including support for compression and avoiding multiple S3 copies. This PR also adds a new buffer type to the S3 sink - Multipart uploads. This PR also includes a number of refactorings to the project and the integration tests.

Signed-off-by: David Venable <dlv@amazon.com>