Updates Kafka configurations such that plugin has its own topic configurations (#3551)
Updates Kafka buffer configurations to disallow configurations which are not needed - e.g. schema and the serde_format for the topic configuration. As part of this change, I also split the TopicConfig into three distinct interfaces and classes. This allows each plugin to either accept a configuration or provide a value of the plugin's own choosing. Also adds copyright headers to all files modified as part of this commit.

Renamed is_topic_create to create_topic. Also made this a boolean internally instead of Boolean since it will have a value.

Adds a zeroBytes() static method to ByteCount as a convenience.

Signed-off-by: David Venable <dlv@amazon.com>