Select require_alias for OS bulk inserts from ISM Policy (#3560)
* Select require_alias for OS bulk inserts from ISM Policy

This change requires an alias when writing to an aliased
index. This avoids creation of an index without alias, when
a previous existing alias and index was deleted. It increases
robustness of DataPrepper's trace index against OS user
interactions.

Signed-off-by: Karsten Schnitter <k.schnitter@sap.com>

* 3342 Determine Alias Configuration from OS

During OS sink initialization it is determined from OS, whether the
configured index actually is an alias. If so, bulk request will require
the index to always be an alias. The response is cached to avoid
further requests. This also ensures, that the alias configuration is
kept in the initially intended state. After all, this change is about to
prevent an automatic index creation for a formerly existing alias.

Signed-off-by: Karsten Schnitter <k.schnitter@sap.com>

* Fix imports for checkstyle

Signed-off-by: Karsten Schnitter <k.schnitter@sap.com>

* Fix integration tests

The specific user used in some tests of OpenSerachSinkIT
needs get permissions on all aliases to test for their existence.
Another bug with determining the alias name is fixed as well.

As a final result, the DataPrepper OpenSearch user requires
write access to the indices and now additionally read access to
the aliases. This can be a change for self-managed indices.

Signed-off-by: Karsten Schnitter <k.schnitter@sap.com>

* Fix Bulk Requests for older OD versions

The `require_alias` parameter for bulk requests was only introduced
with ES 7.10. Since DataPrepper needs to be compatible down to 6.8,
the parameter should not be used in earlier OD versions. This change
will apply the parameter only when OpenSearch is detected as target.

Signed-off-by: Karsten Schnitter <k.schnitter@sap.com>

* Add Permission to get Cluster Info

For checking the OS version, the test user needs an
additional permission.

Signed-off-by: Karsten Schnitter <k.schnitter@sap.com>

---------

Signed-off-by: Karsten Schnitter <k.schnitter@sap.com>