Integration tests for using the Kafka buffer with KMS encryption (#3982)
Adds a new integration test for using the Kafka buffer with KMS encryption. Includes a new KMS CDK stack for any projects that need KMS to use. Some improvements to the CDK stack. Resolves #3980

Change the GitHub tests to include only the Kafka buffer tests which current run in GitHub.

Signed-off-by: David Venable <dlv@amazon.com>