Run tests on the current JVM for Java 17 & 21 / Gradle 8.8 (#4730)
Run tests on the current JVM rather than always using Java 11 for the tests. This fixes a problem with our current GitHub tests where we are running against only Java 11 even though we want to run against different Java versions (11, 17, 21). Updates the Gradle version to 8.8.

Fix Java 21 support in the AbstractSink by removing usage of Thread::stop which now always throws an UnsupportedOperationException.

Use only microsecond precision time when comparing the times in the event_json codec. These tests are failing now on Java 17 and 21 with precision errors.

Fixed a randomly failing test in BlockingBufferTests where a value 0 caused an IllegalArgumentException.

Logging changes to avoid noise in the Gradle builds in GitHub.

Signed-off-by: David Venable <dlv@amazon.com>