Removes the trace-analytics-sample-app from the examples that are provided in the release. (#4867)
The motivation for removing this is that the samples pull in dependencies which often trigger CVE reports. It is not likely customers are trying to run this example from a Data Prepper deployment, especially since the example is not made to run from the installed version, but runs from Docker and runs using the latest 2.x version.

Signed-off-by: David Venable <dlv@amazon.com>
(cherry picked from commit 6ca73cd634fc106cb46defed3d75cd6a73281c99)