Updating the instructions for releasing a new version of Data Prepper.
This adds instructions for the release setup so that maintainers can have a consistent set of instructions to follow. Specifically, it adds steps for setting up the branch, updating version numbers, creating the release notes and change log.

Additionally, this fixes a broken fragment link to the Backporting section of the Developer Guide.

Signed-off-by: David Venable <dlv@amazon.com>