Updates to the code for HTTP chunking. (#4919)
This refactors the code by placing all logic for serializing the data into the Codec itself. In so doing, it allows for improved testing such as symmetric testing. It also decouples the serialization format from the HTTP server. This also uses the Jackson library for the serialization which yields more accurate JSON.

Signed-off-by: David Venable <dlv@amazon.com>