Refactoring chunking in the HTTP source to improve the performance.
The HTTP source was parsing the entire message and then serializing from strings. This created a bit of memory churn and probably duplicate processing. The new approach is to chunk the message from the start which allows us to stream the reading and perform copies of data. This also has a JMH benchmark added which shows that this new approach doubles the number of operations per second.

Signed-off-by: David Venable <dlv@amazon.com>