Fixes and improvements for AbstractSinkTest (#5021)
Fixes and improvements for AbstractSinkTest

Reworked the tests to be somewhat faster. Reworked one flaky test such that it appears to be passing consistently now by using finer-grained control on the timing. Finished one test by making it actually verify the metrics, which is what it was apparently intended to do in the first place.

Be more lenient on the retry thread state. It may be in the sleep() which is ok for the purpose of this test.

Signed-off-by: David Venable <dlv@amazon.com>