Send RetryInfo on OTel Timeouts (#4294)
DataPrepper is sending `RESOURCE_EXHAUSTED` gRPC responses
whenever a buffer is full or a circuit breaker is active. These statuses do
not contain a retry info. In the OpenTelemetry protocol, this implies a
non-retryable error, that will lead to message drops, e.g. in the OTel
collector. To apply proper back pressure in these scenarios a retry info is
added to the status.

Implementation uses exponential backoff. Idea is to start with a minimum
delay on the first time-out or circuit breaker activation. If the next such
event happens within twice the last delay after the previous event, double
the delay until a maximum delay is reached. Use the maximum delay from
then on, until a sufficiently long period (maximum delay) without an event
happens. Then the delay is reset to minimum.

---------

Signed-off-by: Karsten Schnitter <k.schnitter@sap.com>
Signed-off-by: Tomas Longo <tomas.longo@sap.com>
Co-authored-by: David Venable <dlv@amazon.com>