Add traceId to documentId used for spans
adresses #5370

OpenTelemetry spans are currently indexed using the spanId as documentId.
This can lead to collisions where spans are not indexed or overwrite each other.

This change adds the traceId to the documentId. OpenTelemetry assumes the
combination of traceId and spanId to be unique. If there is still a collision it is
safe to assume, that it occurs because of a resending of a previously indexed
span.

Signed-off-by: Karsten Schnitter <k.schnitter@sap.com>