Add integration test forcing the change
Adds an integration tests, that indexes two spans with the same
span id but different trace ids. With the old implementation the two
spans would overwrite each other. This tests checks, that two spans
are returned by the query. This test is only green with the new
document id implementation and fails with the earlier.

Signed-off-by: Karsten Schnitter <k.schnitter@sap.com>