Add traceId to documentId used for spans (#5684)
* Add traceId to documentId used for spans

adresses #5370

OpenTelemetry spans are currently indexed using the spanId as documentId.
This can lead to collisions where spans are not indexed or overwrite each other.

This change adds the traceId to the documentId. OpenTelemetry assumes the
combination of traceId and spanId to be unique. If there is still a collision it is
safe to assume, that it occurs because of a resending of a previously indexed
span.

Signed-off-by: Karsten Schnitter <k.schnitter@sap.com>

* Add integration test forcing the change

Adds an integration tests, that indexes two spans with the same
span id but different trace ids. With the old implementation the two
spans would overwrite each other. This tests checks, that two spans
are returned by the query. This test is only green with the new
document id implementation and fails with the earlier.

Signed-off-by: Karsten Schnitter <k.schnitter@sap.com>

---------

Signed-off-by: Karsten Schnitter <k.schnitter@sap.com>