Updated the smoke tests scripts to use the end-to-end tests (#5903)
Updated the smoke tests scripts to use the end-to-end tests to get them running as part of the release again. Support validating tar files by skipping the Docker pull if the image is local. Skip the Docker validations in the run-smoke-tests.sh script altogether to rely on the e2e-tests. Increase validation time to 30 minutes.

Signed-off-by: David Venable <dlv@amazon.com>