Reduce m365 worker thread count from 7 to 2 (#6103)
This commit reduces the m365 worker thread count from 7 to 2.
This is because the worker thread is per host and we need to ensure
there are not too many threads in total to avoid api throttling.

The suggestion is to estimate based on 2 ~ 4 OCU. Original 7 is the
estimated based on total threads. Changing to 2 would fit for 4 OCU.

Signed-off-by: Wenjie Yao <wjyao@amazon.com>
Co-authored-by: Wenjie Yao <wjyao@amazon.com>