Automatically generate the Data Prepper version from the gradle.properties version.
Updates the approach for getting the current version of Data Prepper to generate a dynamic class in data-prepper-core with the current version of the build. Use Java SPI to load this from DataPrepperVersion in data-prepper-api. With this change we no longer need to update a class file on each version update.

Additionally, because this method is static and used in many other classes, there is now a test library to facilitate getting a version. This one is hard-coded for version 2.

Signed-off-by: David Venable <dlv@amazon.com>