Integration tests for the DynamoDbSourceCoordinationStore (#6190)
Adds integration tests for the DynamoDbSourceCoordinationStore that use an embedded DynamoDB database to test.

To support better testing and cleaner code, this also updates the plugin to use the Spring DI capabilities that Data Prepper now offers to plugin authors.

Exclude the plugin-test-framework project from the DynamoDB source coordination store project because it creates an ANTLR conflict with DynamoDBLocal.

Signed-off-by: David Venable <dlv@amazon.com>