Increase m365 worker thread count from 2 to 4 (#6197)
This commit increases the m365 connector worker thread count from 2 to 4.
This is because the worker thread is per host and we need to ensure
there are not too many threads in total to avoid api throttling.

We had alignment between different teams that 3pE connector will be fixed
at 2 OCU. The performance test confirms a total of 8 threads cross all
fleet. Changing to 4 would fit 2 OCU.

Signed-off-by: Wenjie Yao <wjyao@amazon.com>
Co-authored-by: Wenjie Yao <wjyao@amazon.com>